requirements.txt

markdown==3.5.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-dotenv==1.0.0
pillow==10.2.0